requirements.txt
pandas
appwrite
streamlit
python-dotenv
plotly
flask
geocoder
streamlit_cookies_controller
cryptography